Background-Images for Tracks (#300)
* first draft fetching and using background image

* loading images works

* use device width for fetching images

* use light and dark mode

* save pictures with lightmode/darkmode suffix

* get colormode from settings

* write all backend functions

* polish backend functions

* integrate storage size of pictures into calculation

* deleting track now delete image

* add debug feature to delete all pictures from device

* replace default background images with square images

* add backgroundimage in pictogramm class

* rewrote feedback view

* move responsibility images from tracks to pictogram

* add mapbox attribution logo

* resize image for details view tracks

* rewrite pictogram to match background image

* make backgroundimage use same padding as pictogram to match each other

* remove layoutbuilder from feedbackview

* remove redundant animatedopacity

* resize bottom sheet for track details to better fit image

* pictogram: tweak shadow and reposition mapbox logo

* deleting track now automatically tries to delete image

* remove padding from TrackPainter in Pictogram and put everything in a stack

* use philipps style for map design

* remove string interpolation for formatter

* polish image service

* First working version of the static track background image

* Make extra class for bbox projection code

* Refactor again: prepare for a LRU loop to prune unused images

* Add map to routepictogram

* Add pictograms to shortcuts

* autoprune unused images during app launch

* remove redundant bounding box calculation route_pictogram

* readd total size calculation for images

* readd deletion of associated image when deleting track

* fix padding TrackHistoryItem and move month down

* add padding to backgroundimage

* rebuild feedbackview

* implement Shadowedtext and use it in pictograms

* polish ShadowedText for Pictograms

* move ShadowedText to trackHistoryItems

* add mapbox attribution to images

* make mapbox attribution clickable

* refactore mapbox attribution

* remove redundant aspectratio

* add mapbox style with no text

* routePictogram now uses GPS-data

* switching light and dark mode now changes everything correctly

* use widget height and width for trackHistoryItem instead of hardcoded value

* fix mapbox dialogbox bug

* remove default map image

* Delete redundant pictogram code

* rework track pictograms

* fix track_history pictogram arrangement

* polish borders for trackhistory and shortcuts

* Fix multiple parallel api calls

* fix shadow pictogram in lightmode

* add unique keys for Pictograms

* replace uniquekeys with valuekeys

* remove unused import

* Fix redraw of pictogram in feedback view

* pruneUnusedImages used time now instead of since last fetch

* remove opacity for mapboxlogo

* use pruneUnusedImages in loader with own try-catch-block

* add listener to ShortcutLocationPictogram to react to changing color mode

* last track now reacts to changing colormode

* polish image_cache

* move ShadowedText to text.dart

* redesign pictograms and trackhistoryitems

* delete shadowed text

* remove unused import

* change color text pictorgram darkmode to white

* remind user to restart app after deleting all pictures

* calculate map padding dynamically

* use context for color mode for text in pictogram instead of settings colormode

* add types in map_projection.dart

* add keys to last_track.dart

* fix null exception feedback view

* remove big last ride on home view & manual call for deleting image

* readjust shortcutlocation

* remove redundant opacity = 1

* change wording to Letzte Fahrten

* Add some more niceness

* Add some more niceness v2

* Finish merge of Gamification

* Polishing

* unify color for speed pictogram

* fix shortcuts in gamification & fix setState bug

* replace listeners with didupdatewidget to fix not responding to colormode-change

* fix map_design bug with caused the app not to load

* throw exception when no map design found

* reword exception

* Fix project for iOS 17/Xcode 15

---------

+++ lib/tracking/services/tracking.dart
@@ -532,9 +532,7 @@ class Tracking with ChangeNotifier {
     log.i("Deleting track with id ${track.sessionId}.");
     // Delete the track files.
     final directory = await track.trackDirectory;
-    if (await directory.exists()) {
-      await directory.delete(recursive: true);
-    }
+    if (await directory.exists()) await directory.delete(recursive: true);
     previousTracks?.removeWhere((t) => t.sessionId == track.sessionId);
     await savePreviousTracks();
     notifyListeners();
